More uniform structure of the tests between Scala and Java; reviewed docs
