+++ sqs/src/main/scala/akka/stream/alpakka/sqs/javadsl/SqsPublishSink.scala
@@ -11,7 +11,6 @@ import akka.stream.alpakka.sqs._
 import akka.stream.javadsl.Sink
 import com.amazonaws.services.sqs.AmazonSQSAsync
 import akka.stream.scaladsl.{Flow, Keep}
-import java.lang.{Iterable => JIterable}
 
 import com.amazonaws.services.sqs.model.SendMessageRequest
 
@@ -24,7 +23,7 @@ import scala.compat.java8.FutureConverters.FutureOps
 object SqsPublishSink {
 
   /**
-   * Creates a sink for a SQS queue using an [[com.amazonaws.services.sqs.AmazonSQSAsync]].
+   * Create a sink for a SQS queue accepting Strings.
    */
   def create(queueUrl: String,
              settings: SqsPublishSettings,
@@ -32,13 +31,7 @@ object SqsPublishSink {
     scaladsl.SqsPublishSink.apply(queueUrl, settings)(sqsClient).mapMaterializedValue(_.toJava).asJava
 
   /**
-   * Creates a sink for a SQS queue using an [[com.amazonaws.services.sqs.AmazonSQSAsync]] with default settings.
-   */
-  def create(queueUrl: String, sqsClient: AmazonSQSAsync): Sink[String, CompletionStage[Done]] =
-    create(queueUrl, SqsPublishSettings.Defaults, sqsClient)
-
-  /**
-   * Java API: creates a sink for a SQS queue using an [[com.amazonaws.services.sqs.AmazonSQSAsync AmazonSQSAsync]]
+   * Create a sink for a SQS queue accepting Strings.
    */
   def messageSink(queueUrl: String,
                   settings: SqsPublishSettings,
@@ -49,13 +42,9 @@ object SqsPublishSink {
       .asJava
 
   /**
-   * Java API: creates a sink for a SQS queue using an [[com.amazonaws.services.sqs.AmazonSQSAsync AmazonSQSAsync]] with default settings.
-   */
-  def messageSink(queueUrl: String, sqsClient: AmazonSQSAsync): Sink[SendMessageRequest, CompletionStage[Done]] =
-    messageSink(queueUrl, SqsPublishSettings.Defaults, sqsClient)
-
-  /**
-   * Creates a grouped sink running in batch mode for a SQS queue using an [[com.amazonaws.services.sqs.AmazonSQSAsync]].
+   * Create a flow grouping and publishing `String` messages in batches to a SQS queue.
+   *
+   * @see https://doc.akka.io/docs/akka/current/stream/operators/Source-or-Flow/groupedWithin.html#groupedwithin
    */
   def grouped(queueUrl: String,
               settings: SqsPublishGroupedSettings,
@@ -63,16 +52,9 @@ object SqsPublishSink {
     scaladsl.SqsPublishSink.grouped(queueUrl, settings)(sqsClient).mapMaterializedValue(_.toJava).asJava
 
   /**
-   * Creates a sink running in batch mode for a SQS queue using an [[com.amazonaws.services.sqs.AmazonSQSAsync]] with default settings.
-   */
-  def grouped(queueUrl: String, sqsClient: AmazonSQSAsync): Sink[String, CompletionStage[Done]] =
-    scaladsl.SqsPublishSink
-      .grouped(queueUrl, SqsPublishGroupedSettings.Defaults)(sqsClient)
-      .mapMaterializedValue(_.toJava)
-      .asJava
-
-  /**
-   * Java API: creates a sink running in batch mode for a SQS queue using an [[com.amazonaws.services.sqs.AmazonSQSAsync AmazonSQSAsync]]
+   * Create a flow grouping and publishing `SendMessageRequest` messages in batches to a SQS queue.
+   *
+   * @see https://doc.akka.io/docs/akka/current/stream/operators/Source-or-Flow/groupedWithin.html#groupedwithin
    */
   def groupedMessageSink(queueUrl: String,
                          settings: SqsPublishGroupedSettings,
@@ -83,59 +65,28 @@ object SqsPublishSink {
       .asJava
 
   /**
-   * Java API: creates a sink running in batch mode for a SQS queue using an [[com.amazonaws.services.sqs.AmazonSQSAsync AmazonSQSAsync]] with default settings.
-   */
-  def groupedMessageSink(queueUrl: String, sqsClient: AmazonSQSAsync): Sink[SendMessageRequest, CompletionStage[Done]] =
-    scaladsl.SqsPublishSink
-      .groupedMessageSink(queueUrl, SqsPublishGroupedSettings.Defaults)(sqsClient)
-      .mapMaterializedValue(_.toJava)
-      .asJava
-
-  /**
-   * Creates a sink running in batch mode for a SQS queue using an [[com.amazonaws.services.sqs.AmazonSQSAsync]].
+   * Create a sink running in batch mode for a SQS queue accepting SendMessageRequests.
    */
   def batch(queueUrl: String,
             settings: SqsPublishBatchSettings,
-            sqsClient: AmazonSQSAsync): Sink[JIterable[String], CompletionStage[Done]] =
-    Flow[JIterable[String]]
-      .map(_.asScala.toSeq)
+            sqsClient: AmazonSQSAsync): Sink[java.lang.Iterable[String], CompletionStage[Done]] =
+    Flow[java.lang.Iterable[String]]
+      .map(_.asScala)
       .toMat(scaladsl.SqsPublishSink.batch(queueUrl, settings)(sqsClient))(Keep.right)
       .mapMaterializedValue(_.toJava)
       .asJava
 
   /**
-   * Creates a sink running in batch mode for a SQS queue using an [[com.amazonaws.services.sqs.AmazonSQSAsync]] with default settings.
+   * Create a sink running in batch mode for a SQS queue accepting SendMessageRequests.
    */
-  def batch(queueUrl: String, sqsClient: AmazonSQSAsync): Sink[JIterable[String], CompletionStage[Done]] =
-    Flow[JIterable[String]]
-      .map(jIterable => jIterable.asScala)
-      .toMat(scaladsl.SqsPublishSink.batch(queueUrl, SqsPublishBatchSettings.Defaults)(sqsClient))(Keep.right)
-      .mapMaterializedValue(_.toJava)
-      .asJava
-
-  /**
-   * Java API: creates a sink running in batch mode for a SQS queue using an [[com.amazonaws.services.sqs.AmazonSQSAsync AmazonSQSAsync]]
-   */
-  def batchedMessageSink(queueUrl: String,
-                         settings: SqsPublishBatchSettings,
-                         sqsClient: AmazonSQSAsync): Sink[JIterable[SendMessageRequest], CompletionStage[Done]] =
-    Flow[JIterable[SendMessageRequest]]
-      .map(jIterable => jIterable.asScala)
+  def batchedMessageSink(
+      queueUrl: String,
+      settings: SqsPublishBatchSettings,
+      sqsClient: AmazonSQSAsync
+  ): Sink[java.lang.Iterable[SendMessageRequest], CompletionStage[Done]] =
+    Flow[java.lang.Iterable[SendMessageRequest]]
+      .map(_.asScala)
       .toMat(scaladsl.SqsPublishSink.batchedMessageSink(queueUrl, settings)(sqsClient))(Keep.right)
       .mapMaterializedValue(_.toJava)
       .asJava
-
-  /**
-   * Java API: creates a sink running in batch mode for a SQS queue using an [[com.amazonaws.services.sqs.AmazonSQSAsync AmazonSQSAsync]] with default settings.
-   */
-  def batchedMessageSink(queueUrl: String,
-                         sqsClient: AmazonSQSAsync): Sink[JIterable[SendMessageRequest], CompletionStage[Done]] =
-    Flow[JIterable[SendMessageRequest]]
-      .map(jIterable => jIterable.asScala)
-      .toMat(scaladsl.SqsPublishSink.batchedMessageSink(queueUrl, SqsPublishBatchSettings.Defaults)(sqsClient))(
-        Keep.right
-      )
-      .mapMaterializedValue(_.toJava)
-      .asJava
-
 }
