+++ sqs/src/main/scala/akka/stream/alpakka/sqs/scaladsl/SqsPublishFlow.scala
@@ -5,15 +5,16 @@
 package akka.stream.alpakka.sqs.scaladsl
 
 import akka.NotUsed
-import akka.stream.FlowShape
 import akka.stream.alpakka.sqs.impl.{SqsBatchFlowStage, SqsFlowStage}
 import akka.stream.alpakka.sqs.{
-  SqsPublishResult, SqsPublishBatchSettings, SqsPublishGroupedSettings, SqsPublishSettings}
-import akka.stream.scaladsl.{Flow, GraphDSL}
+  SqsPublishBatchSettings,
+  SqsPublishGroupedSettings,
+  SqsPublishResult,
+  SqsPublishSettings
+}
+import akka.stream.scaladsl.Flow
 import com.amazonaws.services.sqs.AmazonSQSAsync
 import com.amazonaws.services.sqs.model.SendMessageRequest
-
-import scala.concurrent.Future
 
 /**
  * Scala API to create SQS flows.
@@ -21,39 +22,35 @@ import scala.concurrent.Future
 object SqsPublishFlow {
 
   /**
-   * Creates a flow for a SQS queue using an [[com.amazonaws.services.sqs.AmazonSQSAsync]].
+   * Create a flow publishing `SendMessageRequest` messages to a SQS queue.
    */
   def apply(queueUrl: String, settings: SqsPublishSettings = SqsPublishSettings.Defaults)(
       implicit sqsClient: AmazonSQSAsync
   ): Flow[SendMessageRequest, SqsPublishResult, NotUsed] =
-    Flow.fromGraph(new SqsFlowStage(queueUrl, sqsClient)).mapAsync(settings.maxInFlight)(identity)
+    Flow
+      .fromGraph(new SqsFlowStage(queueUrl, sqsClient))
+      .mapAsync(settings.maxInFlight)(identity)
 
+  /**
+   * Create a flow grouping and publishing `SendMessageRequest` messages in batches to a SQS queue.
+   * @see https://doc.akka.io/docs/akka/current/stream/operators/Source-or-Flow/groupedWithin.html#groupedwithin
+   */
   def grouped(queueUrl: String, settings: SqsPublishGroupedSettings = SqsPublishGroupedSettings.Defaults)(
       implicit sqsClient: AmazonSQSAsync
-  ): Flow[SendMessageRequest, SqsPublishResult, NotUsed] = {
-    val graph = GraphDSL.create() { implicit builder =>
-      import GraphDSL.Implicits._
+  ): Flow[SendMessageRequest, SqsPublishResult, NotUsed] =
+    Flow[SendMessageRequest]
+      .groupedWithin(settings.maxBatchSize, settings.maxBatchWait)
+      .via(new SqsBatchFlowStage(queueUrl, sqsClient))
+      .mapAsync(settings.concurrentRequests)(identity)
+      .mapConcat(identity)
 
-      val groupingStage: FlowShape[SendMessageRequest, Seq[SendMessageRequest]] =
-        builder.add(Flow[SendMessageRequest].groupedWithin(settings.maxBatchSize, settings.maxBatchWait))
-
-      val sqsStage: FlowShape[Seq[SendMessageRequest], Future[List[SqsPublishResult]]] =
-        builder.add(new SqsBatchFlowStage(queueUrl, sqsClient))
-
-      val flattenFutures: FlowShape[Future[List[SqsPublishResult]], List[SqsPublishResult]] =
-        builder.add(Flow[Future[List[SqsPublishResult]]].mapAsync(settings.concurrentRequests)(identity))
-
-      val flattenResults: FlowShape[List[SqsPublishResult], SqsPublishResult] = builder.add(Flow[List[SqsPublishResult]].mapConcat(identity))
-      groupingStage ~> sqsStage ~> flattenFutures ~> flattenResults
-
-      FlowShape(groupingStage.in, flattenResults.out)
-    }
-
-    Flow.fromGraph(graph)
-  }
-
+  /**
+   * Create a flow publishing `SendMessageRequest` messages in batches to a SQS queue.
+   */
   def batch(queueUrl: String, settings: SqsPublishBatchSettings = SqsPublishBatchSettings.Defaults)(
       implicit sqsClient: AmazonSQSAsync
   ): Flow[Iterable[SendMessageRequest], List[SqsPublishResult], NotUsed] =
-    Flow.fromGraph(new SqsBatchFlowStage(queueUrl, sqsClient)).mapAsync(settings.concurrentRequests)(identity)
+    Flow
+      .fromGraph(new SqsBatchFlowStage(queueUrl, sqsClient))
+      .mapAsync(settings.concurrentRequests)(identity)
 }
