--- sqs/src/test/scala/akka/stream/alpakka/sqs/scaladsl/SqsAckSettingsSpec.scala
@@ -1,21 +1,0 @@
-/*
- * Copyright (C) 2016-2018 Lightbend Inc. <http://www.lightbend.com>
- */
-
-package akka.stream.alpakka.sqs.scaladsl
-
-import akka.stream.alpakka.sqs.SqsAckSettings
-import org.scalatest.{FlatSpec, Matchers}
-
-class SqsAckSettingsSpec extends FlatSpec with Matchers {
-
-  it should "require valid maxInFlight" in {
-    a[IllegalArgumentException] should be thrownBy {
-      SqsAckSettings(0)
-    }
-  }
-
-  it should "accept valid parameters" in {
-    SqsAckSettings(1)
-  }
-}
